Separate out entangled GP calculation
Why:

Gaussian process fitting was being carried out as part of the init function of
an experiment. This is not necessary, and makes it too hard to modify
the parameters of the GP fit.

The command line utility also lumps together a lot of operations,
making several options too hard to access. I've written two separate utilities
where the GP fit and the MCMC sampling are handled separately.

The old utility (bitc_util.py) is kept for legacy purposes/existing workflows.

This change addresses the need by:

* Writing bitc_integrate.py and bitc_mcmc.py, which handle baseline fitting and mcmc respectively.
* Removing baseline fitting from the init function of ExperimentMicrocal
* Making slight modifications to the bitc_util.py script to handle the changed API