Update mcmc script
Why:

The MCMC script couldn't properly handle competitive experiments.
The command line and the script have been restructured to handle this.
I've also added an example file to test running the two-component model.

Instructions for running the scripts have been added to the readme.

This change addresses the need by:

* Updating readme.md
* Updating the parser for bitc_mcmc.py
* updating the flow inside bitc_mcmc.py